test(iac_scan): #IAC-32 move test to dedicate document
